VFX: the explosion was missing because its shader would not validate
`uMode` is read by both stages of the shell shader. The fragment stage is
mediump and the vertex stage defaults to highp, and GLSL ES requires a uniform
declared in both to carry the same precision in both, so the program linked
and then failed VALIDATE_STATUS — a program that fails validation is not
required to draw anything, and under ANGLE/SwiftShader it draws nothing.

Fireballs, shockwaves, flare cards and scorch decals all share that one
program, so all four pools were dropping out together. That is the actual
reason "explosion.png contains no explosion": the recipe was running, the
sparks were the only part of it with its own shader, and everything else was
never reaching the framebuffer.

Measured on a three-tile sheet at the same frozen camera, before -> after:
110ms mean 130 -> 219, >200 25% -> 76%, warm 11% -> 18%; 340ms warm 8% -> 49%.
It is now far too bright, which is a tuning problem and the next commit's job.

+++ src/gfx/vfx.js
@@ -571,7 +571,15 @@ attribute vec4 aLife;     // birth, life, scale0, scale1
 attribute vec4 aTint;     // rgb + kind
 attribute vec4 aMotion;   // local drift per second + seed
 uniform float uTime;
-uniform float uMode;
+// uMode is read by BOTH stages, and GLSL ES requires a uniform declared in both
+// to carry the same precision in both. The fragment stage is mediump, the
+// vertex stage defaults to highp, so leaving this unqualified linked a program
+// that failed VALIDATE_STATUS — and a program that fails validation is not
+// required to draw anything. That is the actual reason the "explosion" in every
+// capture was a couple of sparks: the fireballs, the shockwaves, the flare
+// cards and the scorch decals all share this one program, so all four pools
+// were silently dropping out together.
+uniform mediump float uMode;
 uniform float uEase;
 varying float vT;
 varying vec4 vTint;
